Update Configurations styles

+++ frontend/src/components/configurations.js
@@ -114,15 +114,18 @@ const Configurations = ({ onFileUpload, jsonData, onSaveConfiguration }) => {
       id
     }
     console.log(idObj)
-    axios.delete(`/api/configurations/${id}`)
-      .then((response) => {
-        console.log('Configuration deleted succesfully:', response.data);
-        setUpdateMsg({msg: response.data.message, color: "#BB86FC"});
-      })
-      .catch((error) => {
-        console.error('Error deleting configuration:', error);
-        setUpdateMsg({msg: error.response.data.error, color: "#CF6679"});
-      });
+    if (window.confirm("Are you sure you want to delete this configuration?"))
+    {
+      axios.delete(`/api/configurations/${id}`)
+        .then((response) => {
+          console.log('Configuration deleted succesfully:', response.data);
+          setUpdateMsg({msg: response.data.message, color: "#BB86FC"});
+        })
+        .catch((error) => {
+          console.error('Error deleting configuration:', error);
+          setUpdateMsg({msg: error.response.data.error, color: "#CF6679"});
+        });
+    } else {setUpdateMsg({msg: "Configuration deletion aborted", color: "#CF6679"})}
   }
 
   return (
@@ -188,14 +191,14 @@ const Configurations = ({ onFileUpload, jsonData, onSaveConfiguration }) => {
             Save Configuration
           </button>
           {(uploadMsg) && (
-            <div style={{color: uploadMsg.color, fontWeight: "bolder", marginTop: "1rem"}}>{uploadMsg.msg}</div>
+            <h2 style={{color: uploadMsg.color, fontWeight: "bolder", marginTop: "1rem"}}>{uploadMsg.msg}</h2>
           )}
         </div>
       )}
       <h2 style={{color: "#BB86FC"}}>--or--</h2>
       <p style={{width: 'fit-content'}}>Select an existing configuration to update or delete:</p>
         {configurations && configurations.length > 0 ? (
-          <div style={{display:'flex'}}>
+          <div className='config-update-container'>
             <div className='select'>
               <select onChange={handleConfigurationSelect}>
                 <option value="select">Select Configuration</option>
@@ -207,50 +210,50 @@ const Configurations = ({ onFileUpload, jsonData, onSaveConfiguration }) => {
               </select>
               <span className='focus'></span>
             </div>
-            {(selectedConfiguration && (selectedConfiguration.name.length > 0)) && (<button onClick={()=>deleteConfiguration(selectedConfiguration._id)}>Delete</button>)}
+            {(selectedConfiguration && (selectedConfiguration.name.length > 0)) && (<button id='delete' onClick={()=>deleteConfiguration(selectedConfiguration._id)}>Delete</button>)}
           </div>
         ) : (
           <p className='no-configs'>No configurations found.</p>
         )}
         {(selectedConfiguration && (selectedConfiguration.name.length > 0)) && (
           <div className="column-mapping">
-          <h2>Column Mapping:</h2>
-          <table>
-            <thead>
-              <tr>
-                <th>Column</th>
-                <th>Value</th>
-              </tr>
-            </thead>
-            <tbody>
-              {Object.keys(selectedConfiguration.columnMappings).map((columnName, index) => (
-                <tr key={index}>
-                  <td>{columnName}</td>
-                  <td className='select'>
-                    <select
-                      name={index}
-                      value={columnMappings[index]}
-                      onChange={handleColumnMapping}
-                    >
-                      <option value="">Select Value</option>
-                      <option value="fullName">Full Name</option>
-                      <option value="phoneNumber">Phone Number</option>
-                      <option value="emailAddress">Email Address</option>
-                      <option value="party">Party</option>
-                      <option value="age">Age</option>
-                    </select>
-                  </td>
+            <h2>Column Mapping:</h2>
+            <table>
+              <thead>
+                <tr>
+                  <th>Column</th>
+                  <th>Value</th>
                 </tr>
-              ))}
-            </tbody>
-          </table>
-          <input type="text" value={name} onChange={handleInputChange} placeholder="Configuration Name" />
-          <button onClick={() => updateConfiguration(name, selectedConfiguration._id, columnMappings)}>
-            Update Configuration
-          </button>
-          {(updateMsg) && (
-            <div style={{color: updateMsg.color, fontWeight: "bolder", marginTop: "1rem"}}>{updateMsg.msg}</div>
-          )}
+              </thead>
+              <tbody>
+                {Object.keys(selectedConfiguration.columnMappings).map((columnName, index) => (
+                  <tr key={index}>
+                    <td>{columnName}</td>
+                    <td className='select'>
+                      <select
+                        name={index}
+                        value={columnMappings[index]}
+                        onChange={handleColumnMapping}
+                      >
+                        <option value="">Select Value</option>
+                        <option value="fullName">Full Name</option>
+                        <option value="phoneNumber">Phone Number</option>
+                        <option value="emailAddress">Email Address</option>
+                        <option value="party">Party</option>
+                        <option value="age">Age</option>
+                      </select>
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+            <input type="text" value={name} onChange={handleInputChange} placeholder="Configuration Name" />
+            <button id='update' onClick={() => updateConfiguration(name, selectedConfiguration._id, columnMappings)}>
+              Update Configuration
+            </button>
+            {(updateMsg) && (
+              <h2 style={{color: updateMsg.color, fontWeight: "bolder", marginTop: "1rem"}}>{updateMsg.msg}</h2>
+            )}
         </div>
         )}
     </div>
